@ style(blind-token): satisfy current nightly rustfmt + clippy
Current floating-nightly rustfmt reformats several blind-token sources
(perturb.rs, scheme.rs, wire/mod.rs) and clippy flags an unnecessary
i64->i64 cast in ring.rs mont_reduce. Formatting-only + the redundant
cast removed; no behavioural change (KATs unchanged).
@

+++ lib-q-blind-token/src/lattice/ring.rs
@@ -80,7 +80,7 @@ fn mont_reduce(t: u128) -> i64 {
     // Branchless conditional subtract: r - (q if r >= q else 0).
     let d = r - Q; // ∈ [-q, q)
     let mask = d >> 63; // all-ones iff d < 0 (i.e. r < q), else 0
-    (d + (mask & Q)) as i64
+    d + (mask & Q)
 }
 
 /// `a·b mod q` for `a, b ∈ [0, q)`, via two Montgomery reductions (division-free; far faster than
